refactor: split lock-harness into inspect/link/batch modules
Pure code move: path/status inspection, mutation primitives, and bulk
operations each get their own module; lock-harness.ts stays as a
re-export barrel so all existing imports keep working.
